refactor: static DAO 삭제
