+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,60 +1,72 @@
 package wooteco.subway.dao;
 
-import org.springframework.util.ReflectionUtils;
+import org.springframework.dao.support.DataAccessUtils;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Station;
 
-import java.lang.reflect.Field;
-import java.util.ArrayList;
+import java.sql.PreparedStatement;
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public class LineDao {
-    private static final String ERROR_MESSAGE_NOT_FOUND_LINE_ID = "Id에 해당하는 노선이 없습니다.";
-    private static final long AFFECTED_ROWS_COUNT = 1;
+    private JdbcTemplate jdbcTemplate;
 
-    private static Long seq = 0L;
-    private static List<Line> lines = new ArrayList<>();
-
-    public static Line save(String name, String color) {
-        Line line = new Line(++seq, name, color);
-        lines.add(line);
-        return line;
+    public LineDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
-    private static Station createNewObject(Station station) {
-        Field field = ReflectionUtils.findField(Station.class, "id");
-        field.setAccessible(true);
-        ReflectionUtils.setField(field, station, ++seq);
-        return station;
+    private final RowMapper<Line> lineRowMapper = (resultSet, rowNum) ->
+            new Line(
+                    resultSet.getLong("id"),
+                    resultSet.getString("name"),
+                    resultSet.getString("color"));
+
+    public Line save(String name, String color) {
+        String sql = "INSERT INTO LINE (name, color) VALUES (?, ?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, name);
+            ps.setString(2, color);
+            return ps;
+        }, keyHolder);
+        Long createdLineId = keyHolder.getKey().longValue();
+        return new Line(createdLineId, name, color);
     }
 
-    public static List<Line> findAll() {
-        return lines;
+    public List<Line> findAll() {
+        String query = "SELECT * FROM LINE";
+        return jdbcTemplate.query(query, lineRowMapper);
     }
 
-    public static void deleteAll() {
-        lines.clear();
+    public Optional<Line> findById(Long lineId) {
+        String query = "SELECT * FROM LINE WHERE id = ?";
+        Line result = DataAccessUtils.singleResult(
+                jdbcTemplate.query(query, lineRowMapper, lineId)
+        );
+        return Optional.ofNullable(result);
     }
 
-    public static Optional<Line> findById(Long lineId) {
-        return lines.stream()
-                .filter(line -> line.getId().equals(lineId))
-                .findFirst();
+    public Optional<Line> findByName(String name) {
+        String query = "SELECT * FROM LINE WHERE name = ?";
+        Line result = DataAccessUtils.singleResult(
+                jdbcTemplate.query(query, lineRowMapper, name)
+        );
+        return Optional.ofNullable(result);
     }
 
-    public static Long edit(Long lineId, String color, String name) {
-        Line foundLine = findById(lineId)
-                .orElseThrow(() -> new IllegalArgumentException(ERROR_MESSAGE_NOT_FOUND_LINE_ID));
-        int index = lines.indexOf(foundLine);
-        lines.set(index, new Line(lineId, color, name));
-        return AFFECTED_ROWS_COUNT;
+    public Long edit(Long lineId, String color, String name) {
+        String query = "UPDATE LINE SET color = ?, name = ? WHERE id = ?";
+        return (long) jdbcTemplate.update(query, color, name, lineId);
     }
 
-    public static Long deleteById(Long lineId) {
-        Line foundLine = findById(lineId)
-                .orElseThrow(() -> new IllegalArgumentException(ERROR_MESSAGE_NOT_FOUND_LINE_ID));
-        lines.remove(foundLine);
-        return AFFECTED_ROWS_COUNT;
+    public Long deleteById(Long lineId) {
+        String query = "DELETE FROM LINE WHERE id = ?";
+        return (long) jdbcTemplate.update(query, lineId);
     }
 }
