+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,59 +1,57 @@
 package wooteco.subway.dao;
 
-import org.springframework.util.ReflectionUtils;
+import org.springframework.dao.support.DataAccessUtils;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
 import wooteco.subway.domain.Station;
 
-import java.lang.reflect.Field;
-import java.util.ArrayList;
+import java.sql.PreparedStatement;
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public class StationDao {
-    private static final String ERROR_MESSAGE_NOT_FOUND_LINE_ID = "Id에 해당하는 노선이 없습니다.";
-    private static final long AFFECTED_ROWS_COUNT = 1;
-    private static Long seq = 0L;
-    private static List<Station> stations = new ArrayList<>();
+    private JdbcTemplate jdbcTemplate;
 
-    public static Station save(Station station) {
-        validateDuplicate(station);
-        Station persistStation = createNewObject(station);
-        stations.add(persistStation);
-        return persistStation;
+    public StationDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
-    private static void validateDuplicate(Station station) {
-        boolean isDuplicate = stations.stream()
-                .anyMatch(s -> s.getName().equals(station.getName()));
-        if (isDuplicate) {
-            throw new IllegalArgumentException("이미 존재하는 이름입니다.");
-        }
+    private final RowMapper<Station> stationRowMapper = (resultSet, rowNum) ->
+            new Station(
+                    resultSet.getLong("id"),
+                    resultSet.getString("name"));
+
+    public Station save(String name) {
+        String sql = "INSERT INTO STATION (name) VALUES (?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, name);
+            return ps;
+        }, keyHolder);
+        Long createdStationId = keyHolder.getKey().longValue();
+        return new Station(createdStationId, name);
     }
 
-    public static List<Station> findAll() {
-        return stations;
+    public Optional<Station> findByName(String name) {
+        String query = "SELECT * FROM STATION WHERE name = ?";
+        Station result = DataAccessUtils.singleResult(
+                jdbcTemplate.query(query, stationRowMapper, name)
+        );
+        return Optional.ofNullable(result);
     }
 
-    private static Station createNewObject(Station station) {
-        Field field = ReflectionUtils.findField(Station.class, "id");
-        field.setAccessible(true);
-        ReflectionUtils.setField(field, station, ++seq);
-        return station;
+    public List<Station> findAll() {
+        String query = "SELECT * FROM STATION";
+        return jdbcTemplate.query(query, stationRowMapper);
     }
 
-    public static Optional<Station> findById(Long id) {
-        return stations.stream()
-                .filter(station -> station.getId().equals(id))
-                .findFirst();
-    }
-
-    public static Long deleteById(Long id) {
-        Station foundStation = findById(id)
-                .orElseThrow(() -> new IllegalArgumentException(ERROR_MESSAGE_NOT_FOUND_LINE_ID));
-        stations.remove(foundStation);
-        return AFFECTED_ROWS_COUNT;
-    }
-
-    public static void deleteAll() {
-        stations.clear();
+    public Long deleteById(Long stationId) {
+        String query = "DELETE FROM STATION WHERE id = ?";
+        return (long) jdbcTemplate.update(query, stationId);
     }
 }
